+++ lib/geordi/commands/update.rb
@@ -10,13 +10,24 @@ option :dump, type: :string, aliases: '-d', banner: 'TARGET',
 option :test, type: :boolean, aliases: '-t', desc: 'After updating, run tests'
 
 def update
+  old_ruby_version = File.read('.ruby-version').chomp
+
   Interaction.announce 'Updating repository'
   Util.run!('git pull', show_cmd: true)
 
-  invoke_geordi 'migrate'
+  ruby_version = File.read('.ruby-version').chomp
+  ruby_version_changed = !ruby_version.empty? && (ruby_version != old_ruby_version)
 
-  Interaction.success 'Successfully updated the project.'
+  if ruby_version_changed
+    puts
+    Interaction.warn 'Ruby version changed during git pull. Please run again to use the new version.'
+    exit(1)
+  else
+    invoke_geordi 'migrate'
 
-  invoke_geordi 'dump', options.dump, load: true if options.dump
-  invoke_geordi 'tests' if options.test
+    Interaction.success 'Successfully updated the project.'
+
+    invoke_geordi 'dump', options.dump, load: true if options.dump
+    invoke_geordi 'tests' if options.test
+  end
 end